In test.py, the timigs are now written to timings.pickle.arch file, where arch is the architecture of the machine where the test was run.
The changes in c-code treat now the Numeric arrays as long integers,
so that the code works also in architecture where long and int are of
different size.



git-svn-id: https://svn.fysik.dtu.dk/projects/gpaw/trunk@76 6419b465-d902-0410-997c-be2266baf723

+++ c/bc.h
@@ -38,8 +38,8 @@ typedef struct
 const static int COPY_DATA = -2;
 const static int DO_NOTHING = -3; // ??????????
 
-boundary_conditions* bc_init(const int size1[3], const int padding[2], 
-			     const int neighbors[3][2],
+boundary_conditions* bc_init(const long size1[3], const int padding[2], 
+			     const long neighbors[3][2],
 			     MPI_Comm comm, bool real, bool cfd);
 void bc_unpack1(const boundary_conditions* bc, 
 		const double* input, double* output, int i,
@@ -53,4 +53,4 @@ void bc_unpack2(const boundary_conditions* bc,
 		MPI_Request sendreq[2], 
 		double* rbuf);
 void bc_set_rotation(boundary_conditions* bc,
-		     double angle, double* coefs, int* offsets, int exact);
+		     double angle, double* coefs, long* offsets, int exact);
